Changed convert binary

+++ app/src/main/java/com/example/myapplication/domain/usecase/ConvertToDecUseCase.kt
@@ -4,7 +4,7 @@ import kotlin.math.pow
 
 class ConvertToDecUseCase {
 
-    fun execute(value: String, numberSystem: Double): String {
+    fun execute(value: String, numberSystem: Int): String {
         return if (value.contains(".")) {
             val ratio = value.split(".")
             (convertToDec(ratio[0], numberSystem, false) +
@@ -13,7 +13,7 @@ class ConvertToDecUseCase {
             convertToDec(value, numberSystem, false).toString()
     }
 
-    private fun convertToDec(value: String, numberSystem: Double, valAfterDot: Boolean): Double {
+    private fun convertToDec(value: String, numberSystem: Int, valAfterDot: Boolean): Double {
         var decResult = 0.0
         var index = 0
         val size = value.length - 1
@@ -22,8 +22,7 @@ class ConvertToDecUseCase {
             index = value.length
 
         for (it in value) {
-            decResult += numberSystem.pow(size - index) * it.digitToInt()
-
+            decResult += numberSystem.toDouble().pow((size - index).toDouble()) * it.digitToInt()
             index++
         }
